Add API functionality
Implemented API endpoints for retrieving menu data and changing menu status. Also added a trait for generating API responses, and refactored existing code to utilize the trait. Modified error messages for when data is not found. Resolved issue where an error was not displayed when data was not sent. Modified response format to include a dynamically generated status code. Changed the response format to XML.
i removed xml (just for learn